feat: datapack-driven worldgen, /datapack & /locate biome commands, entity metadata fixes
Datapack-driven worldgen:
- Embed the vanilla worldgen JSON baseline (density_function/ and noise/ assets) via build.rs codegen, so datapacks are resolved against the same registry state real Minecraft starts from.
- Add a runtime flattener (datapack/flatten.rs) that lowers parsed density-function DAGs into the flat BaseNoiseFunctionComponent stack the noise interpreter consumes, including splines, wrappers, and interpolated noise.
- Fix noise-override precedence: a datapack that ships its own worldgen/noise/<id>.json for a minecraft: key (e.g. tectonic's continentalness) now re-seeds that noise instead of being ignored. Biome placement now matches a datapack-only reference world at 99.92%.
- Add runtime multi-noise biome placement for datapack biomes (biome_placement.rs, biome_source codec) and a surface_rule codec that maps noise_settings surface rules onto the generated MaterialRule types.
- Extract md5_to_lo_hi in xoroshiro128 to seed noises from registry keys the way vanilla does; extend the noise router and proto_chunk/generator wiring to accept runtime-built routers

Datapack selection (level.dat parity):
- Read DataPacks.Enabled/Disabled from level.dat and feed worldgen sources in enabled order (later = higher priority, last-write-wins
- Auto-enable packs present in <world>/datapacks that aren't explicitly disabled

Commands:
- /datapack with vanilla-parity list/enable/disable output.
- /locate biome with a biome resource argument (spiral scan out to 6400 blocks, 64-block steps), including dynamic datapack biomes.
- /seed now displays the seed as a signed i64 like vanilla.

Entity fixes:
- Regenerate 26.1/26.2 tracked data: wolf VARIANT_ID moved from index 21 (colliding with COLLAR_COLOR and disconnecting clients with "Invalid entity data item type for field 21") to its correct index 23; send the wolf variant metadata and add a regression test.
- Track each entity's persisted chunk (first_loaded_chunk_position) so moving entities are removed from their old chunk on save instead of being duplicated on reload.

Cleanup:
- Remove the superseded single-file data_component_impl.rs (replaced by the data_component_impl/ module directory), regenerate chunk_gen_settings and noise_router data, refresh Cargo.lock.

+++ pumpkin/src/command/commands/seed.rs
@@ -37,7 +37,7 @@ fn create_copy_on_click_text(content: String) -> TextComponent {
 impl CommandExecutor for SeedCommandExecutor {
     fn execute<'a>(&'a self, context: &'a CommandContext) -> CommandExecutorResult<'a> {
         Box::pin(async move {
-            let seed = context.world().level.seed.0;
+            let seed = context.world().level.seed.0 as i64;
             let seed_string = seed.to_string();
 
             context
